fix (rewrite) the query heap
Note that there is potentially for the heap to leak since we are
capturing the top-level environment. Although we restrict the
environment to the functions used by calls in the expression, those
functions might be enclosed in an environment that references
the parent query of a later deferred translation (that references the
current one through its query on the heap). If that happens,
the later deferred translation will never be freed, and so neither will
the earlier one.